+++ src/Core2D/Layout/PointUtil.cs
@@ -5,15 +5,15 @@ namespace Core2D.Layout
 {
     internal static class PointUtil
     {
-        public static double Snap(double value, double snap)
+        public static decimal Snap(decimal value, decimal snap)
         {
-            if (snap == 0.0)
+            if (snap == 0m)
             {
                 return value;
             }
-            decimal c = (decimal)value % (decimal)snap;
-            decimal r = c >= (decimal)snap / 2m ? (decimal)value + (decimal)snap - c : (decimal)value - c;
-            return (double)r;
+            decimal c = value % snap;
+            decimal r = c >= snap / 2m ? value + snap - c : value - c;
+            return r;
         }
 
         public static int CompareX(IPointShape point1, IPointShape point2)
@@ -26,10 +26,10 @@ namespace Core2D.Layout
             return (point1.Y > point2.Y) ? 1 : ((point1.Y < point2.Y) ? -1 : 0);
         }
 
-        public static void Rotate(IPointShape point, double radians, double centerX, double centerY, out double x, out double y)
+        public static void Rotate(IPointShape point, decimal radians, decimal centerX, decimal centerY, out decimal x, out decimal y)
         {
-            x = (point.X - centerX) * Math.Cos(radians) - (point.Y - centerY) * Math.Sin(radians) + centerX;
-            y = (point.X - centerX) * Math.Sin(radians) + (point.Y - centerY) * Math.Cos(radians) + centerY;
+            x = ((decimal)point.X - centerX) * (decimal)Math.Cos((double)radians) - ((decimal)point.Y - centerY) * (decimal)Math.Sin((double)radians) + centerX;
+            y = ((decimal)point.X - centerX) * (decimal)Math.Sin((double)radians) + ((decimal)point.Y - centerY) * (decimal)Math.Cos((double)radians) + centerY;
         }
     }
 }
