+++ src/Core2D/ViewModels/Shapes/GroupShape.cs
@@ -106,7 +106,7 @@ namespace Core2D.Shapes
         }
 
         /// <inheritdoc/>
-        public override void Move(ISelection selection, double dx, double dy)
+        public override void Move(ISelection selection, decimal dx, decimal dy)
         {
             foreach (var shape in Shapes)
             {
